Scaffold function

+++ Makefile
@@ -2,6 +2,7 @@ TEMPLATES_DIR := templates
 TEMPLATE_EXTENSION := yaml
 PARAMETERS_DIR := parameters
 PARAMETER_EXTENSION := json
+FUNCTIONS_DIR := lambdas
 
 
 
@@ -16,7 +17,47 @@ provision:
 	  --parameters file://$(PARAMETERS_DIR)/$(PARAMS).$(PARAMETER_EXTENSION) \
 	  --capabilities CAPABILITY_IAM
 
-
 delete:
+	test -n "$(STACK_NAME)" || { echo "No STACK_NAME given!"; exit 1; }
+
 	aws cloudformation delete-stack \
 	  --stack-name $(STACK_NAME)
+
+define INDEX_BODY
+import logging
+import os
+
+# from $(FUNCTION) import *
+
+logging.basicConfig()
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+
+def handler(event, context):
+    """Entry point for the Lambda function."""
+    logger.info('Hello world')
+
+
+if __name__ == '__main__':
+    handler(None, None)
+endef
+export INDEX_BODY
+
+foo:
+	@echo "$$INDEX_BODY" > foo.txt
+
+scaffold:
+	test -n "$(FUNCTION)" || { echo "No FUNCTION given!"; exit 1; }
+	test ! -d $(FUNCTIONS_DIR)/$(FUNCTION)_package || { echo "$(FUNCTION)_package already exists!"; exit 1; }
+
+	mkdir -p $(FUNCTIONS_DIR)/$(FUNCTION)_package
+	mkdir -p $(FUNCTIONS_DIR)/$(FUNCTION)_package/requirements
+	echo "pyaml" >> $(FUNCTIONS_DIR)/$(FUNCTION)_package/requirements/common.txt
+	echo "-r common.txt" >> $(FUNCTIONS_DIR)/$(FUNCTION)_package/requirements/dev.txt
+	echo "pytest" >> $(FUNCTIONS_DIR)/$(FUNCTION)_package/requirements/dev.txt
+	echo "mock" >> $(FUNCTIONS_DIR)/$(FUNCTION)_package/requirements/dev.txt
+	echo "-r common.txt" >> $(FUNCTIONS_DIR)/$(FUNCTION)_package/requirements/lambda.txt
+	mkdir -p $(FUNCTIONS_DIR)/$(FUNCTION)_package/$(FUNCTION)
+	echo "$$INDEX_BODY" > $(FUNCTIONS_DIR)/$(FUNCTION)_package/index.py
+
